+++ src/web/assets/ckeditor/src/ckeditor5-craftcms.js
@@ -415,6 +415,8 @@ const handleClipboard = function (editor, plugins) {
         let duplicatedContent = pasteContent;
         let errors = false;
         const siteId = Craft.siteId;
+        let ownerId = null;
+        let layoutElementUid = null;
         const editorData = editor.getData();
         const matches = [...pasteContent.matchAll(/data-entry-id="([0-9]+)/g)];
 
@@ -422,6 +424,21 @@ const handleClipboard = function (editor, plugins) {
         // we need to get duplicates and update the content snippet that's being pasted in
         // before we can call further events
         event.stop();
+
+        const $editorElement = $(editor.ui.view.element);
+        const $parentForm = $editorElement.parents('form');
+        let elementEditor = $parentForm.data('elementEditor');
+
+        // ensure we're working with a draft
+        await elementEditor.ensureIsDraftOrRevision();
+
+        // get the target owner id, in case we're pasting to a different element all together
+        ownerId = elementEditor.settings.elementId;
+
+        // get the target field id, in case we're pasting to a different field all together (not different instance, different field)
+        layoutElementUid = $editorElement
+          .parents('.field')
+          .data('layoutElement');
 
         // for each nested entry ID we found
         for (let i = 0; i < matches.length; i++) {
@@ -466,6 +483,8 @@ const handleClipboard = function (editor, plugins) {
                     entryId: entryId,
                     siteId: siteId,
                     targetEntryTypeIds: targetEntryTypeIds,
+                    targetOwnerId: ownerId,
+                    targetLayoutElementUid: layoutElementUid,
                   },
                 },
               )