+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - CKEditor fields can now be merged together. ([#277](https://github.com/craftcms/ckeditor/pull/277))
 - Added `craft\ckeditor\migrations\BaseConvertMatrixContentMigration`.
 - Fixed a bug where CKEditor fields’ search keywords were including nested entries’ rendered partial templates rather than nested entries’ search keywords.
+- Fixed a bug where nested entries weren’t always getting propagated correctly. ([#254](https://github.com/craftcms/ckeditor/issues/254))
 
 ## 4.1.0 - 2024-06-12